Move Date types and functions into date.go; even more flexible parsing
Need to strip any extra whitespace before passing dates to time.Parse to get
valid dates out of it.

IMAP servers will probably respond with many variations of the
2-Jan-2006 15:04:05 -0700 date layout.

Not sure if it's best to strip whitespace in the Date/Time functions, or
to do it before data gets to the Date/Time functions.