imap: quote parameter values
An email from outlook.com contained an embedded part with
Content-ID and filename starting with a long sequence of
digits:

	279351849188E3D84D6DCC5B@namprd05.prod.outlook.com

When sent unquoted, the Apple Mail IMAP client attempts to
parse this initially as an integer. When it overflows a
uint32 parsing is stopped, so the message contents are never
displayed.

Quoting all parameter values is safe and avoids this.
Indeed the IMAP RFC includes examples of far more aggressive
quoting of almost all strings from a MIME message.
That may be worth considering in the future