Write NIL for empty ENVELOPE fields
RFC 3501 requires NIL for missing fields and "" for empty fields,
however, acknowledges the fact some servers may return NIL for empty
fields. We do not bother complicating the structure and do the same.

It is believed that empty-but-present header fields are rare enough
to be ignored, while returning "" for missing fields is more of an
issue. The main reason for this change is to silence imaptest, though.