v2: Extend backend interface to support compatible overlapping extensions
See GH#322 for related discussion. This commit takes a slightly
different approach for interface design by using slices of opaque
values instead of a struct. This has a number of benefits, notably
backend can easily detect that it receives an "option" that it
cannot handle while in struct that would not be possible to
see new fields. This, however, creates an similar situation
but for go-imap. It is believed that it is not a significant
issue since we expect many backends and one go-imap :)

Interface changes are mostly applied to command affected by
RFC 4466 which defines an uniform framework for command
extensions.

In this design, it is important for go-imap to know the exact
set of extensions supported by backend beforehand (e.g.
which Options it can handle). SupportedExtensions() is
added to accommodate that. For consistency purposes, it should
also be used for extensions that can be easily checked using
type assertions (e.g. extensions adding brand new commands).