Fix FETCH command with ALL, FAST and FULL macro as fetch items
When using one of the macros as fetch item arguments, the IMAP command
written to the socket would wrap the macro in parentheses, which is not
valid.

To address this we check whether there is only one item in the fetch
list and if it matches on of the macros we make sure it's serialized in
way that does not wrap it in parentheses.