imapclient/search: handle UID SEARCH results without hits
When searching for unseen mails we can get the following error if there
are none.

> in response-data: imapwire: expected SP, got "\r"

The reason is that the IMAP server is sending a response like this
in case there are no search hits.

T7 UID SEARCH RETURN (ALL) (UNSEEN)
* ESEARCH (TAG "T7") UID
T7 OK SEARCH completed (Success)

We can see that there is no SP after "UID".

As emersion correctly points out, the ABNF grammar for ESEARCH[0] accepts
replies that don't even contain the "UID".

We change the parser to accept such replies as well.

Suggested-by: emersion

[0] https://www.rfc-editor.org/rfc/rfc4466#section-3